Add the remaining linear and postal symbologies with a physical size model
Codabar, Code 11, MSI, Plessey, Telepen, Channel Code, PosiCode, BC412,
Pharmacode, Two-track Pharmacode, Flattermarken, FIM and a raw linear
symbology. POSTNET, PLANET, Royal Mail 4-state, KIX, DAFT, Intelligent
Mail, Australia Post, Japan Post and 4-state Mailmark.

BarcodeOptions sizes the symbol in millimetres: an X dimension and a bar
height, rendered at a Dpi, with the nominal values of each specification
as the defaults. A text placement names the edge its ink starts on, and
each encoder adds the clear space its specification gives. Symbols whose
bars are one width at one pitch draw every run at a whole number of
pixels.

+++ src/ImageSharp.Drawing/Barcodes/BarcodeSymbol.cs
@@ -23,12 +23,13 @@ internal abstract class BarcodeSymbol
 
     /// <summary>
     /// Gets the width in modules of the quiet zone that comes before the symbol. Each symbology sets its
-    /// own quiet zone width. For example, ISO/IEC 15420 requires 11 modules before an EAN-13 symbol.
+    /// own quiet zone width. For example, ISO/IEC 15420 requires 11 modules before an EAN-13 symbol, and
+    /// a postal symbology states its clear zone in inches, which is not a whole number of modules.
     /// </summary>
-    public abstract int LeadingQuietZone { get; }
+    public abstract float LeadingQuietZone { get; }
 
     /// <summary>
     /// Gets the width of the quiet zone that must follow the symbol, in modules.
     /// </summary>
-    public abstract int TrailingQuietZone { get; }
+    public abstract float TrailingQuietZone { get; }
 }
