+++ tests/ImageSharp.Drawing.Tests/Barcodes/Code39SymbologyTests.cs
@@ -7,8 +7,8 @@ namespace SixLabors.ImageSharp.Drawing.Tests.Barcodes;
 
 /// <summary>
 /// Encoder tests for <see cref="Code39Symbology"/>. Each expected run string is the alternating bar and
-/// space module widths, starting with a bar, taken from an independent reference implementation through
-/// its raw encoding API, less the gap the reference emits after the stop character. Section 4.2 of
+/// space module widths, starting with a bar, from the reference implementation's raw encoding API, less
+/// the gap the reference emits after the stop character. Section 4.2 of
 /// ISO/IEC 16388 separates the characters "within the symbol" with that gap, and section 4.3.3 puts the
 /// stop character at the right end, so the runs end on a bar.
 /// </summary>
@@ -177,7 +177,7 @@ public class Code39SymbologyTests
     public void MatchesTheSymbolWidthFormula(string text, CheckCharacterMode checkCharacter, int characters)
     {
         LinearBarcodeSymbol symbol = Encode(new Code39Symbology(checkCharacter), text);
-        int drawn = symbol.LeadingQuietZone + symbol.TrailingQuietZone;
+        float drawn = symbol.LeadingQuietZone + symbol.TrailingQuietZone;
         foreach (int run in symbol.RunWidths)
         {
             drawn += run;
